에피소드 등록 화면 구현

+++ src/main/java/com/wtoon/webtoon/model/dto/Webtoon.java
@@ -34,7 +34,7 @@ public class Webtoon {
 	private List<String> hashtag;
 	private List<String> day; //연재요일
 	private List<String> genre;
-	private int avgRating;	//평균별점
+	private double avgRating;	//평균별점
 	
     // 요일 번호와 요일 이름 매핑
     private static final Map<String, String> DAY_MAP = Stream.of(new String[][] {
